requirements.txt
opencv-python
numpy
ipykernel
tensorflow
ultralytics
Pillow
supervision
numpy
streamlit
tensorflow
keras

# absl-py==2.1.0
# altair==5.4.1
# asttokens==2.4.1
# astunparse==1.6.3
# attrs==24.2.0
# blinker==1.9.0
# cachetools==5.5.0
# certifi==2024.8.30
# charset-normalizer==3.4.0
# click==8.1.7
# colorama==0.4.6
# comm==0.2.2
# contourpy==1.3.0
# cycler==0.12.1
# debugpy==1.8.8
# decorator==5.1.1
# defusedxml==0.7.1
# executing==2.1.0
# filelock==3.16.1
# flatbuffers==24.3.25
# fonttools==4.54.1
# fsspec==2024.10.0
# gast==0.6.0
# gitdb==4.0.11
# GitPython==3.1.43
# google-pasta==0.2.0
# grpcio==1.67.1
# h5py==3.12.1
# idna==3.10
# ipykernel==6.29.5
# ipython==8.29.0
# jedi==0.19.2
# Jinja2==3.1.4
# jsonschema==4.23.0
# jsonschema-specifications==2024.10.1
# jupyter_client==8.6.3
# jupyter_core==5.7.2
# keras==3.6.0
# kiwisolver==1.4.7
# libclang==18.1.1
# Markdown==3.7
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# matplotlib==3.9.2
# matplotlib-inline==0.1.7
# mdurl==0.1.2
# ml-dtypes==0.4.1
# mpmath==1.3.0
# namex==0.0.8
# narwhals==1.13.3
# nest-asyncio==1.6.0
# networkx==3.4.2
# numpy==2.0.2
# opencv-python==4.10.0.84
# opt_einsum==3.4.0
# optree==0.13.0
# packaging==24.2
# pandas==2.2.3
# parso==0.8.4
# pillow==11.0.0
# platformdirs==4.3.6
# prompt_toolkit==3.0.48
# protobuf==5.28.3
# psutil==6.1.0
# pure_eval==0.2.3
# py-cpuinfo==9.0.0
# pyarrow==18.0.0
# pydeck==0.9.1
# Pygments==2.18.0
# pyparsing==3.2.0
# python-dateutil==2.9.0.post0
# pytz==2024.2
# pywin32==308
# PyYAML==6.0.2
# pyzmq==26.2.0
# referencing==0.35.1
# requests==2.32.3
# rich==13.9.4
# rpds-py==0.21.0
# scipy==1.14.1
# seaborn==0.13.2
# setuptools==75.3.0
# six==1.16.0
# smmap==5.0.1
# stack-data==0.6.3
# streamlit==1.40.0
# supervision==0.24.0
# sympy==1.13.1
# tenacity==9.0.0
# tensorboard==2.18.0
# tensorboard-data-server==0.7.2
# tensorflow==2.18.0
# tensorflow_intel==2.18.0
# termcolor==2.5.0
# toml==0.10.2
# torch==2.5.1
# torchvision==0.20.1
# tornado==6.4.1
# tqdm==4.67.0
# traitlets==5.14.3
# typing_extensions==4.12.2
# tzdata==2024.2
# ultralytics==8.3.28
# ultralytics-thop==2.0.11
# urllib3==2.2.3
# watchdog==5.0.3
# wcwidth==0.2.13
# Werkzeug==3.1.3
# wheel==0.45.0
# wrapt==1.16.0